# backend-node/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: 宽窄·Orbit 业务后端 API
  description: |
    宽窄·Orbit 社区型 APP 业务后端 API 契约。

    ## 认证
    使用 Bearer Token 方式：`Authorization: Bearer <access_token>`

    ## 统一响应格式
    ```json
    { "success": true, "data": { ... } }
    ```
    错误时：
    ```json
    { "success": false, "error": { "code": "40001", "message": "..." }, "requestId": "req_xxx" }
    ```

    ## 错误码
    | Code | 含义 |
    |------|------|
    | 40001 | 请求参数错误 |
    | 40101 | 未登录或 Token 失效 |
    | 40301 | 无权限或被限制 |
    | 40401 | 资源不存在 |
    | 40901 | 状态冲突 / 重复操作 / 额度不足 |
    | 42901 | 请求频率过高 |
    | 50000 | 服务内部错误 |
    | 50401 | 上游超时 |

    ## 幂等
    写操作推荐携带 `Idempotency-Key` 请求头（UUID），相同 Key 的重复请求返回相同结果。

    ## 分页
    使用 page-based 分页：`?page=1&pageSize=20`，响应包含 `hasMore` 和 `nextPage`。
  version: 1.0.0
  contact:
    name: 宽窄·Orbit Backend Team
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.kuanzhai-orbit.com/api/v1
    description: Production
...
paths:
...
  /activities/list:
    get:
      tags: [activities]
      summary: 获取活动列表
      operationId: getActivities
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 活动列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: '#/components/schemas/Activity'
                          hasMore:
                            type: boolean
                          nextPage:
                            type: integer
  /activities/{id}:
    get:
      tags: [activities]
      summary: 获取活动详情
      operationId: getActivityDetail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 活动详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Activity'
  /activities/{id}/join:
    post:
      tags: [activities]
      summary: 报名活动
      operationId: joinActivity
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 报名成功（幂等）
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          status:
                            type: string
                            enum: [pending, approved, waitlist]
  /activities/{id}/join-status:
    get:
      tags: [activities]
      summary: 获取报名状态
      operationId: getJoinStatus
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 报名状态
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          status:
                            type: string
                            enum: [none, pending, approved, waitlist, rejected]

  # ─────────────── Share ───────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
...
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
...
    Activity:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        status:
          type: string
          enum: [registering, ongoing, ended]
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        capacity:
          type: integer
        participantCount:
          type: integer
        joinStatus:
          type: string
          enum: [none, pending, approved, waitlist, rejected]
...